Add CreateTeamOption
